+++ src/include/StateMachine.hpp
@@ -13,10 +13,10 @@ namespace targetfinder {
         friend class Marker;
     public:
         static constexpr int MIN_LENGTH = 1;
-        static constexpr double TOLERANCE = 0.5;
+        static constexpr float TOLERANCE = 0.5;
         static constexpr int NUM_STATES = 7;
 
-        StateMachine(int input_length=0, double tolerance=TOLERANCE, int min_pixels=MIN_LENGTH);
+        StateMachine(int input_length=0, float tolerance=TOLERANCE, int min_pixels=MIN_LENGTH);
         Marker* step(int x, int y, bool value1, bool value2=true);
         void reset(int y, bool value1=false, bool value2=true);
         cv::Vec3b state_colour();
@@ -32,7 +32,7 @@ namespace targetfinder {
         int first_x, last_x, y, input_length, min_pixels, max_pixels;
         int min_bound, max_bound;
         int pixel_counts[NUM_STATES];
-        double tolerance;
+        float tolerance;
 
 
     };
