+++ src/utils/Utils.h
@@ -12,24 +12,24 @@
 #define MIN(a, b) ((a < b) ? a : b)
 #define MAX(a, b) ((a > b) ? a : b)
 
-static double _angle(double ax, double ay, double bx, double by) {
+static float _angle(float ax, float ay, float bx, float by) {
     if(bx - ax == 0 || by - ay == 0) {
         return 0.0;
     }
     return atan2(bx - ax, by - ay);
 }
 
-static double _degrees(double radians) {
+static float _degrees(float radians) {
     return (radians * 180.0) / M_PI;
 }
 
-static double _aspect(double a1, double a2) {
-    double a = MAX(a1, a2);
-    double b = MIN(a1, a2);
+static float _aspect(float a1, float a2) {
+    float a = MAX(a1, a2);
+    float b = MIN(a1, a2);
     return b / a;
 }
 
-static double _distance(double ax, double bx, double ay, double by) {
+static float _distance(float ax, float bx, float ay, float by) {
     return sqrt(
             pow(bx - ax, 2) + pow(by - ay, 2)
     );
